@@ -5,7 +5,8 @@ import android.graphics.drawable.Icon;
 import android.os.Build;
 import android.service.quicksettings.Tile;
 import android.service.quicksettings.TileService;
-import android.support.annotation.RequiresApi;
+
+import androidx.annotation.RequiresApi;
 
 import javax.inject.Inject;
 
