@@ -1,8 +1,9 @@
 package co.garmax.materialflashlight.ui;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
 import android.view.View;
+
+import androidx.fragment.app.Fragment;
 
 import dagger.android.support.AndroidSupportInjection;
 
