@@ -1,7 +1,8 @@
 package co.garmax.materialflashlight.ui;
 
-import android.support.annotation.NonNull;
 import android.text.TextUtils;
+
+import androidx.annotation.NonNull;
 
 import org.junit.BeforeClass;
 
